+++ cmd/pomerium/main.go
@@ -4,7 +4,7 @@ import (
 	"flag"
 	"fmt"
 	"net/http"
-	"os"
+	"sync"
 	"time"
 
 	"github.com/fsnotify/fsnotify"
@@ -15,6 +15,7 @@ import (
 	"github.com/pomerium/pomerium/authenticate"
 	"github.com/pomerium/pomerium/authorize"
 	"github.com/pomerium/pomerium/internal/config"
+	"github.com/pomerium/pomerium/internal/grpcutil"
 	"github.com/pomerium/pomerium/internal/httputil"
 	"github.com/pomerium/pomerium/internal/log"
 	"github.com/pomerium/pomerium/internal/middleware"
@@ -30,36 +31,41 @@ var versionFlag = flag.Bool("version", false, "prints the version")
 var configFile = flag.String("config", "", "Specify configuration file location")
 
 func main() {
+	if err := run(); err != nil {
+		log.Fatal().Err(err).Msg("cmd/pomerium")
+	}
+}
+
+func run() error {
 	flag.Parse()
 	if *versionFlag {
 		fmt.Println(version.FullVersion())
-		os.Exit(0)
+		return nil
 	}
 	opt, err := config.ParseOptions(*configFile)
 	if err != nil {
-		log.Fatal().Err(err).Msg("cmd/pomerium: options")
+		return err
 	}
 	log.Info().Str("version", version.FullVersion()).Msg("cmd/pomerium")
-
-	setupMetrics(opt)
+	// since we can have multiple listeners, we create a wait group
+	var wg sync.WaitGroup
+	setupMetrics(opt, &wg)
 	setupTracing(opt)
-	setupHTTPRedirectServer(opt)
+	setupHTTPRedirectServer(opt, &wg)
 
 	r := newGlobalRouter(opt)
-	grpcServer := setupGRPCServer(opt)
-	_, err = newAuthenticateService(*opt, r.Host(urlutil.StripPort(opt.AuthenticateURL.Host)).Subrouter())
+	_, err = newAuthenticateService(*opt, r)
 	if err != nil {
-		log.Fatal().Err(err).Msg("cmd/pomerium: authenticate")
+		return err
 	}
-
-	authz, err := newAuthorizeService(*opt, grpcServer)
+	authz, err := newAuthorizeService(*opt, &wg)
 	if err != nil {
-		log.Fatal().Err(err).Msg("cmd/pomerium: authorize")
+		return err
 	}
 
 	proxy, err := newProxyService(*opt, r)
 	if err != nil {
-		log.Fatal().Err(err).Msg("cmd/pomerium: proxy")
+		return err
 	}
 	if proxy != nil {
 		defer proxy.AuthorizeClient.Close()
@@ -71,13 +77,15 @@ func main() {
 		log.Info().Str("file", e.Name).Msg("cmd/pomerium: config file changed")
 		opt = config.HandleConfigUpdate(*configFile, opt, []config.OptionsUpdater{authz, proxy})
 	})
-	srv, err := httputil.NewTLSServer(configToServerOptions(opt), r, grpcServer)
-	if err != nil {
-		log.Fatal().Err(err).Msg("cmd/pomerium: couldn't start pomerium")
-	}
-	httputil.Shutdown(srv)
 
-	os.Exit(0)
+	srv, err := httputil.NewServer(httpServerOptions(opt), r, &wg)
+	if err != nil {
+		return err
+	}
+	go httputil.Shutdown(srv)
+	// Blocks and waits until ALL WaitGroup members have signaled completion
+	wg.Wait()
+	return nil
 }
 
 func newAuthenticateService(opt config.Options, r *mux.Router) (*authenticate.Authenticate, error) {
@@ -88,11 +96,13 @@ func newAuthenticateService(opt config.Options, r *mux.Router) (*authenticate.Au
 	if err != nil {
 		return nil, err
 	}
-	r.PathPrefix("/").Handler(service.Handler())
+	sr := r.Host(urlutil.StripPort(opt.AuthenticateURL.Host)).Subrouter()
+	sr.PathPrefix("/").Handler(service.Handler())
+
 	return service, nil
 }
 
-func newAuthorizeService(opt config.Options, rpc *grpc.Server) (*authorize.Authorize, error) {
+func newAuthorizeService(opt config.Options, wg *sync.WaitGroup) (*authorize.Authorize, error) {
 	if !config.IsAuthorize(opt.Services) {
 		return nil, nil
 	}
@@ -100,7 +110,18 @@ func newAuthorizeService(opt config.Options, rpc *grpc.Server) (*authorize.Autho
 	if err != nil {
 		return nil, err
 	}
-	pbAuthorize.RegisterAuthorizerServer(rpc, service)
+	regFn := func(s *grpc.Server) {
+		pbAuthorize.RegisterAuthorizerServer(s, service)
+	}
+	so := &grpcutil.ServerOptions{
+		Addr:      opt.GRPCAddr,
+		SharedKey: opt.SharedKey,
+	}
+	if !opt.GRPCInsecure {
+		so.TLSCertificate = opt.TLSCertificate
+	}
+	grpcSrv := grpcutil.NewServer(so, regFn, wg)
+	go grpcutil.Shutdown(grpcSrv)
 	return service, nil
 }
 
@@ -146,21 +167,7 @@ func newGlobalRouter(o *config.Options) *mux.Router {
 	return mux
 }
 
-func configToServerOptions(opt *config.Options) *httputil.ServerOptions {
-	return &httputil.ServerOptions{
-		Addr:              opt.Addr,
-		Cert:              opt.Cert,
-		Key:               opt.Key,
-		CertFile:          opt.CertFile,
-		KeyFile:           opt.KeyFile,
-		ReadTimeout:       opt.ReadTimeout,
-		WriteTimeout:      opt.WriteTimeout,
-		ReadHeaderTimeout: opt.ReadHeaderTimeout,
-		IdleTimeout:       opt.IdleTimeout,
-	}
-}
-
-func setupMetrics(opt *config.Options) {
+func setupMetrics(opt *config.Options, wg *sync.WaitGroup) {
 	if opt.MetricsAddr != "" {
 		if handler, err := metrics.PrometheusHandler(); err != nil {
 			log.Error().Err(err).Msg("cmd/pomerium: metrics failed to start")
@@ -168,7 +175,7 @@ func setupMetrics(opt *config.Options) {
 			metrics.SetBuildInfo(opt.Services)
 			metrics.RegisterInfoMetrics()
 			serverOpts := &httputil.ServerOptions{Addr: opt.MetricsAddr}
-			srv := httputil.NewHTTPServer(serverOpts, handler)
+			srv, _ := httputil.NewServer(serverOpts, handler, wg)
 			go httputil.Shutdown(srv)
 		}
 	}
@@ -191,18 +198,21 @@ func setupTracing(opt *config.Options) {
 	}
 }
 
-func setupHTTPRedirectServer(opt *config.Options) {
+func setupHTTPRedirectServer(opt *config.Options, wg *sync.WaitGroup) {
 	if opt.HTTPRedirectAddr != "" {
 		serverOpts := httputil.ServerOptions{Addr: opt.HTTPRedirectAddr}
-		srv := httputil.NewHTTPServer(&serverOpts, httputil.RedirectHandler())
+		srv, _ := httputil.NewServer(&serverOpts, httputil.RedirectHandler(), wg)
 		go httputil.Shutdown(srv)
 	}
 }
 
-func setupGRPCServer(opt *config.Options) *grpc.Server {
-	grpcAuth := middleware.NewSharedSecretCred(opt.SharedKey)
-	grpcOpts := []grpc.ServerOption{
-		grpc.UnaryInterceptor(grpcAuth.ValidateRequest),
-		grpc.StatsHandler(metrics.NewGRPCServerStatsHandler(opt.Services))}
-	return grpc.NewServer(grpcOpts...)
+func httpServerOptions(opt *config.Options) *httputil.ServerOptions {
+	return &httputil.ServerOptions{
+		Addr:              opt.Addr,
+		TLSCertificate:    opt.TLSCertificate,
+		ReadTimeout:       opt.ReadTimeout,
+		WriteTimeout:      opt.WriteTimeout,
+		ReadHeaderTimeout: opt.ReadHeaderTimeout,
+		IdleTimeout:       opt.IdleTimeout,
+	}
 }